requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx